+++ cpp09/ex01/RPN.hpp
@@ -4,23 +4,21 @@
 #include"iostream"
 #include<stack>
 
-//bool isOperand(char c);
 bool validExpr(std::string& input);
-//int operation(std::stack<int> *rpn, char o, int x);
 void execExpr(std::string& expr);
 
 class Operation{
-	int a;
-	int b;
+	long long a;
+	long long b;
 	char sign;
 
 	public:
-	Operation(int a, int b, char s);
+	Operation(long long a, long long b, char s);
 	Operation(const Operation& other);
 	Operation& operator=(const Operation& other);
 	~Operation();
 
-	int operation();
+	long long operation();
 	static bool isOperand(char c);
 };
